+++ changelog.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [1.19.4 - 1.1.0]
+## [1.20 - 1.1.0]
 
-* port: 1.19.4
+* port: 1.20
 
 ## [Fix]
