+++ gradle.properties
@@ -8,10 +8,10 @@ mod_group=com.troblecodings
 mod_version=1.1.0
 
 # Minecraft
-minecraft_version=1.19.4
+minecraft_version=1.20
 
 # Forge
-forge_version=45.4.3
+forge_version=46.0.14
 loader=forge
 
 # Java
