+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockGhostGlowstone.java
@@ -2,12 +2,10 @@ package com.troblecodings.invisiblelights.blocks;
 
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
-import net.minecraft.world.level.material.Material;
 
 public class BlockGhostGlowstone extends Block {
 
     public BlockGhostGlowstone() {
-        super(Block.Properties.of(Material.GLASS).sound(SoundType.GLASS).strength(0.3f)
-                .noOcclusion());
+        super(Block.Properties.of().sound(SoundType.GLASS).strength(0.3f).noOcclusion());
     }
 }
