+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockInvisibleLight.java
@@ -5,7 +5,6 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.RenderShape;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.material.Material;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
@@ -13,8 +12,7 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 public class BlockInvisibleLight extends Block {
 
     public BlockInvisibleLight(final int light) {
-        super(Block.Properties.of(Material.DIRT).lightLevel(state -> light).strength(0.5f)
-                .noOcclusion());
+        super(Block.Properties.of().lightLevel(state -> light).strength(0.5f).noOcclusion());
     }
 
     @Override
